Hopefully fixed 2 errors:
Coefficient of bin, negbin distributions was approximated in the past.
Now we replaced it by a pre-computed table of log(ints).  Factorial can
be built as log(N!) = log(N) + log(N-1) + …

Also, was causing problems when switching a trial to a binomial
distribution, but the counts of that trial were > n of the binomial.